+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - `simpleprometheusreceiver`: Add support for static labels (#7908)
 - `podmanreceiver`: Add container.runtime attribute to container metrics (#8262)
 - `dockerstatsreceiver`: Add container.runtime attribute to container metrics (#8261)
+- `clickhouseexporter`: Implement consume log logic. (#9705)
 
 ### 🛑 Breaking changes 🛑